Implement authorization using Policy

+++ app/Http/Controllers/QuestionController.php
@@ -7,6 +7,11 @@ use App\Models\Question;
 
 class QuestionController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index', 'show');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -56,11 +61,13 @@ class QuestionController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Question  $question
+     * @param  \App\Models\Question $question
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Question $question)
     {
+        $this->authorize('update', $question);
         return view('questions.edit')->with(['question' => $question]);
     }
 
@@ -82,10 +89,11 @@ class QuestionController extends Controller
      *
      * @param  \App\Models\Question $question
      * @return \Illuminate\Http\RedirectResponse
-     * @throws \Exception
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Question $question)
     {
+        $this->authorize('delete', $question);
         $question->delete();
         return redirect()->route('questions.index')->with(['success' => 'Your question has been deleted']);
     }
